[Configs] Refactor configs into different files.
