+++ config/src/config/utils.rs
@@ -1,0 +1,36 @@
+// Copyright © Aptos Foundation
+// SPDX-License-Identifier: Apache-2.0
+
+use std::path::{Path, PathBuf};
+
+#[derive(Debug)]
+pub struct RootPath {
+    root_path: PathBuf,
+}
+
+impl RootPath {
+    pub fn new<P: AsRef<Path>>(path: P) -> Self {
+        let root_path = if let Some(parent) = path.as_ref().parent() {
+            parent.to_path_buf()
+        } else {
+            PathBuf::from("")
+        };
+
+        Self { root_path }
+    }
+
+    /// This function assumes that the path is already a directory
+    pub fn new_path<P: AsRef<Path>>(path: P) -> Self {
+        let root_path = path.as_ref().to_path_buf();
+        Self { root_path }
+    }
+
+    /// This adds a full path when loading / storing if one is not specified
+    pub fn full_path(&self, file_path: &Path) -> PathBuf {
+        if file_path.is_relative() {
+            self.root_path.join(file_path)
+        } else {
+            file_path.to_path_buf()
+        }
+    }
+}
